Changes in accessibility
In AccessibleMainMenuActivity.java :
1. When talkback mode on, and then menu is pressed then back, the
talkback mode turns off. This happens when accessibility is on too, but
shouldn't happen.
Rectified by launching talkback mode on back button press, and removing
"Read to me"  from accessible menu.

In SpeakActivity.java :
2. When accessibility mode is on and screen locks, the voice stops. When
screen unlocked the talkback is off, but "Pause" displayed.

This bug removed, and it is shown opaque, and label changed to "Play".